requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
crypto==1.4.1
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.9
Django==3.2.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
msgpack==1.0.2
Naked==0.1.31
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
service-identity==21.1.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
Twisted==21.2.0
websocket-client==1.0.1
